requirements.txt
streamlit>=1.28.0
ccxt>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-telegram-bot>=20.0
